Refactor size to be int instead of float

+++ Assets/Scripts/CameraSystem.cs
@@ -41,7 +41,7 @@ public class CameraSystem : ComponentSystem
         float progress = 10 * Time.deltaTime;
 
         _mainCamera.transform.position = Vector3.Lerp(_mainCamera.transform.position, targetPos, progress);
-        float sizeRatio = m_Data.Size[0].Value / settings.PlayerInitialSize;
+        float sizeRatio = (float) m_Data.Size[0].Value / settings.PlayerInitialSize;
         float newCameraSize = math.sqrt(sizeRatio) * settings.InitialCameraSize;
         _mainCamera.orthographicSize = math.lerp(_mainCamera.orthographicSize, newCameraSize, progress);
     }
